+++ CHANGELOG.md
@@ -19,7 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Add `key()`, `start()`, `value()` to `StoreKeyStartValue`
  - Add `StorageError::MissingMetadata` now that implicit groups are not supported
  - Add `ChunkShape::to_array_shape()`
- - Add `DataTypeMetadata`
+ - Add `DataTypeMetadataV3`
 
 ### Changed
  - **Breaking**: `Arc` instead of `Box` partial decoders
@@ -42,9 +42,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
         - `ZARR_NAN_{F16, BF16, F32, F64}`
         - `ChunkKeySeparator`
         - `DimensionName`
-    - Split `DataType` into `array::DataType` and `metadata::_::DataTypeMetadata`
+    - Split `DataType` into `array::DataType` and `metadata::_::DataTypeMetadataV3`
  - **Breaking**: `data_key` and `[async_]retrieve_partial_values` in `zarrs::storage` now take a `chunk_key` instead of a `chunk_key_encoding` and `chunk_indices`
- - **Breaking**: Move `metadata::{v2,v3}::{codec,chunk_grid,chunk_key_encoding}` to `metadata::{v2,v3}::array::`
+ - **Breaking**: Move `metadata::{v2,v3}::{codec,chunk_grid,chunk_key_encoding}` to `metadata::{v2,v3}::array::
+ - **Breaking**: Rename `ArrayMetadataV2DataType` to `DataTypeMetadataV2`
 
 ### Removed
  - **Breaking**: Remove `array::NonZeroError`, use `std::num::TryFromIntError` instead